profile routes now check requesting user

+++ src/main/java/com/codecool/binder/service/ProfileService.java
@@ -2,17 +2,21 @@ package com.codecool.binder.service;
 
 import com.codecool.binder.dto.ProfileDto;
 import com.codecool.binder.model.Profile;
+import com.codecool.binder.model.User;
 import com.codecool.binder.repository.ProfileRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ProfileService {
     private final ProfileRepository repository;
+    private final UserService userService;
 
     @Autowired
-    public ProfileService(ProfileRepository repository) {
+    public ProfileService(ProfileRepository repository, UserService userService) {
         this.repository = repository;
+        this.userService = userService;
     }
 
     public ProfileDto convert (Profile profile) {
@@ -24,16 +28,34 @@ public class ProfileService {
                 .build();
     }
 
-    public ProfileDto saveProfile (Profile profile) {
+    public ProfileDto getProfileDto (Long id, String sessionUserEmail) {
+        User user = userService.getUserByEmail(sessionUserEmail);
+        Profile profile = repository.getOne(id);
+        if (profile.getOwner().equals(user) || profile.getOwner().isMatched(user)) {
+            return convert(profile);
+        } else throw new BadCredentialsException("Access denied.");
+    }
+
+    public ProfileDto createProfile (Profile profile, String sessionUserEmail) {
+        User user = userService.getUserByEmail(sessionUserEmail);
+        profile.setOwner(user);
+        profile.setId(null);
         repository.save(profile);
-        return convert(profile);
+        return convert(repository.getOne(profile.getId()));
     }
 
-    public ProfileDto getProfileDto(Long id) {
-        return convert(repository.getOne(id));
+    public ProfileDto updateProfile (Profile profile, String sessionUserEmail) {
+        User user = userService.getUserByEmail(sessionUserEmail);
+        if (profile.getOwner().equals(user)) {
+            repository.save(profile);
+            return convert(repository.getOne(profile.getId()));
+        } else throw new BadCredentialsException("Access denied.");
     }
 
-    public void deleteProfile(Long id) {
-        repository.deleteById(id);
+    public void deleteProfile (Long id, String sessionUserEmail) {
+        User user = userService.getUserByEmail(sessionUserEmail);
+        if (repository.getOne(id).getOwner().equals(user)) {
+            repository.deleteById(id);
+        } else throw new BadCredentialsException("Access denied.");
     }
 }
